requirements.txt
pymupdf>=1.24.0
python-docx>=1.1.0
pyobjc-framework-Vision>=10.0
pyobjc-framework-Quartz>=10.0
